requirements.txt
antlr4-python3-runtime==4.7.2
astroid==2.3.2
autopep8==1.4.4
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
coverage==5.1
crayons==0.3.0
cryptography==2.9.2
Django==2.2.10
django-admin-tools==0.8.1
django-appconf==1.0.3
django-classy-tags==0.9.0
django-compressor==2.3
django-cookie-law==2.0.2
django-extensions==2.2.5
django-mathfilters==1.0.0
-e git+https://github.com/dprog-philippe-docourt/django-qr-code.git@d8811b7e05b1beff466c7800007b603c3c1eae7f#egg=django_qr_code
django-sass-processor==0.7.4
django-simple-cookie-consent==0.1.1
djangorestframework==3.10.3
elastos-adenine==2.0.2
ExtensionClass==4.4
fastecdsa==1.7.5
grpcio==1.28.1
grpcio-tools==1.28.1
gunicorn==20.0.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
libsass==0.19.4
mccabe==0.6.1
namegenerator==1.0.6
Pillow==6.2.2
protobuf==3.11.1
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.3
pytest-runner==5.2
python-decouple==3.3
pytz==2020.1
qrcode==6.1
rcssmin==1.0.6
Record==3.5
requests==2.23.0
rjsmin==1.1.0
rollbar==0.15.0
selenium==3.141.0
six==1.12.0
solidity-parser==0.0.7
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.6
webdriver-manager==2.5.1
wrapt==1.11.2
